+++ data/src/main/java/in/koreatech/koin/data/source/remote/RecruitmentRemoteDataSource.kt
@@ -1,6 +1,7 @@
 package `in`.koreatech.koin.data.source.remote
 
 import `in`.koreatech.koin.data.api.auth.RecruitmentAuthApi
+import `in`.koreatech.koin.data.response.recruitment.MyRecruitmentListResponse
 import javax.inject.Inject
 import retrofit2.HttpException
 
@@ -29,6 +30,18 @@ class RecruitmentRemoteDataSource @Inject constructor(
 
     suspend fun readAllNotifications() {
         val response = recruitmentAuthApi.readAllNotifications()
+        if (!response.isSuccessful) throw HttpException(response)
+    }
+
+    suspend fun getMyRecruitmentPosts(
+        status: String,
+        sort: String,
+        page: Int,
+        limit: Int
+    ): MyRecruitmentListResponse = recruitmentAuthApi.getMyRecruitmentPosts(status, sort, page, limit)
+
+    suspend fun closeRecruitmentPost(postId: Int) {
+        val response = recruitmentAuthApi.closeRecruitmentPost(postId)
         if (!response.isSuccessful) throw HttpException(response)
     }
 }
