+++ domain/src/main/java/in/koreatech/koin/domain/repository/RecruitmentRepository.kt
@@ -1,5 +1,6 @@
 package `in`.koreatech.koin.domain.repository
 
+import `in`.koreatech.koin.domain.model.recruitment.MyRecruitmentPost
 import `in`.koreatech.koin.domain.model.recruitment.RecruitmentNotifications
 
 interface RecruitmentRepository {
@@ -10,4 +11,13 @@ interface RecruitmentRepository {
     suspend fun readNotification(notificationId: Int): Result<Unit>
 
     suspend fun readAllNotifications(): Result<Unit>
+
+    suspend fun getMyRecruitmentPosts(
+        status: String,
+        sort: String,
+        page: Int,
+        limit: Int
+    ): Result<List<MyRecruitmentPost>>
+
+    suspend fun closeRecruitmentPost(postId: Int): Result<Unit>
 }
