--- domain/src/main/java/in/koreatech/koin/domain/repository/TeamRecruitmentRepository.kt
@@ -1,15 +1,0 @@
-package `in`.koreatech.koin.domain.repository
-
-import `in`.koreatech.koin.domain.model.recruitment.MyRecruitmentPost
-
-interface TeamRecruitmentRepository {
-
-    suspend fun getMyRecruitmentPosts(
-        status: String,
-        sort: String,
-        page: Int,
-        limit: Int
-    ): Result<List<MyRecruitmentPost>>
-
-    suspend fun closeRecruitmentPost(postId: Int): Result<Unit>
-}
