+++ domain/src/main/java/in/koreatech/koin/domain/usecase/recruitment/CloseRecruitmentPostUseCase.kt
@@ -1,11 +1,11 @@
 package `in`.koreatech.koin.domain.usecase.recruitment
 
-import `in`.koreatech.koin.domain.repository.TeamRecruitmentRepository
+import `in`.koreatech.koin.domain.repository.RecruitmentRepository
 import javax.inject.Inject
 
 class CloseRecruitmentPostUseCase @Inject constructor(
-    private val teamRecruitmentRepository: TeamRecruitmentRepository
+    private val recruitmentRepository: RecruitmentRepository
 ) {
     suspend operator fun invoke(postId: Int): Result<Unit> =
-        teamRecruitmentRepository.closeRecruitmentPost(postId)
+        recruitmentRepository.closeRecruitmentPost(postId)
 }
