+++ feature/recruitment/src/main/java/in/koreatech/koin/feature/recruitment/navigation/RecruitmentNavType.kt
@@ -27,7 +27,7 @@ sealed class RecruitmentNavType {
     data class ApplicantManagement(val postId: Int) : RecruitmentNavType()
 
     @Serializable
-    data class ApplicantDetail(val postId: Long, val applicantId: Long) : RecruitmentNavType()
+    data class ApplicantDetail(val postId: Int, val applicantId: Int) : RecruitmentNavType()
 
     @Serializable
     data object MyRecruitment : RecruitmentNavType()
