+++ feature/recruitment/src/main/java/in/koreatech/koin/feature/recruitment/ui/applicantdetail/ApplicantDetailScreen.kt
@@ -168,7 +168,7 @@ private fun ApplicantDetailScreenPreview() {
     RebrandKoinTheme {
         ApplicantDetailScreenImpl(
             applicant = ApplicantDetail(
-                id = 1L,
+                id = 1,
                 name = "김철수",
                 role = "프론트엔드",
                 department = "컴퓨터공학부",
